Changed to long. IT WORKED!!!

+++ NTT.h
@@ -4,6 +4,8 @@
 
 #ifndef NTT_NTT_H
 #define NTT_NTT_H
+
+
 #include "params.h"
 
 
@@ -17,7 +19,7 @@
  * @param start The level where it is started
  * @param stop  How many levels the splitting should be
  */
-void initiate_NTT_forward(int *roots, int *NTT_forward, int n, int move, bool direction , int start, int  stop);
+void initiate_NTT_forward(long *roots, long *NTT_forward, long n, long move, bool direction , long start, long  stop);
 
 /**
  * Creates list of roots that the multiplication will do modulo when doing multiplication after NTTforward
@@ -25,10 +27,10 @@ void initiate_NTT_forward(int *roots, int *NTT_forward, int n, int move, bool di
  * @param level How many levels that will be computed
  * @param NTT_roots The list where the roots will be stored
  */
-void initiate_NTT_roots(int *NTT_forward, int level, int *NTT_roots);
+void initiate_NTT_roots(long *NTT_forward, long level, long *NTT_roots);
 
-int NTT_forward[NUM_POLYNOMIALS*2];
-int NTT_roots[NUM_POLYNOMIALS];
+long NTT_forward[NUM_POLYNOMIALS*2];
+long NTT_roots[NUM_POLYNOMIALS];
 
 /**
  * Computes the forward NTT. This doesn´t work as of 8.11.20
@@ -43,7 +45,7 @@ int NTT_roots[NUM_POLYNOMIALS];
  * @author Anna Bakkebø
  * @date 8.11.20
  */
-void forward_NTT(int *pol, int i,int  n,int m, int * roots, int start, int stop);
+void forward_NTT(long *pol, long i,long  n,long m, long * roots, long start, long stop);
 
 /**
  * Computes forward NTT
@@ -54,7 +56,7 @@ void forward_NTT(int *pol, int i,int  n,int m, int * roots, int start, int stop)
  * @param levels  how many levels the NTT will be performed
  * @param n  The length of the polynomial that will be splitted
  */
-void forward_NTT2(int *pol, int *NTT_forward,int move, int start, int levels, int n);
+void forward_NTT2(long *pol, long *NTT_forward,long move, long start, long levels, long n);
 
 /**
  * Computes inverse NTT
@@ -66,14 +68,14 @@ void forward_NTT2(int *pol, int *NTT_forward,int move, int start, int levels, in
  * @param levels how many levels that will be performed
  * @param n Twice the length of each of the small polynomial
  */
-void innverse_NTT2(int *pol, int *NTT_forward,int move, int start, int levels, int n);
+void innverse_NTT2(long *pol, long *NTT_forward,long move, long start, long levels, long n);
 void inverse_NTT();
 
 /**
  * Multiplying by the innverse of 2^LEVEL
  * @param pol The polynomial right after the innverse_NTT2 to finnish the inverse algorithm
  */
-void innverse_finnish(int *pol);
+void innverse_finnish(long *pol);
 
 /**
  * printing the NTTroots variable
