+++ params.c
@@ -7,12 +7,12 @@
 
 
 
-void load_roots(int *roots){
-    int w = 1;
-    int q = Q;
+void load_roots(long *roots){
+    long w = 1;
+    long q = Q;
     //printf("q=%d\n",q);
     roots[0]=1;
-    for (int i = 1; i<PRIMITIVE_N; i++){
+    for (long i = 1; i<PRIMITIVE_N; i++){
         w = (w*ROOT_OF_UNITY %q);
         roots[i]=w;
         /*w = w*ROOT_OF_UNITY %q;
@@ -22,7 +22,7 @@ void load_roots(int *roots){
 
 void print_roots(){
     printf("The roots of are: {");
-    for(int i=0;i<PRIMITIVE_N;i++){
+    for(long i=0;i<PRIMITIVE_N;i++){
         printf("%d, ",roots[i]);
     }
     printf("}\n");
